Moved Weapon's offset into it's construct

+++ Bullet-Inferno/src/se/dat255/bulletinferno/model/entity/DefaultBossImpl.java
@@ -39,7 +39,7 @@ public class DefaultBossImpl extends SimpleBoss implements Ship {
 	 */
 	public DefaultBossImpl(PhysicsEnvironment physics, EntityEnvironment entities, EnemyDefinitionImpl type, 
 
-			Vector2 position, Vector2 velocity, PhysicsMovementPattern pattern, int initialHealth, 
+			Vector2 position, Vector2 velocity, PhysicsMovementPattern pattern, float initialHealth, 
 
 			Weapon[] weapons, int score, int credits, 
 
@@ -56,7 +56,7 @@ public class DefaultBossImpl extends SimpleBoss implements Ship {
 	public DefaultBossImpl(PhysicsEnvironment physics, EntityEnvironment entities, EnemyDefinitionImpl type, 
 
 
-			Vector2 position, Vector2 velocity, int initialHealth, Weapon[] weapons, 
+			Vector2 position, Vector2 velocity, float initialHealth, Weapon[] weapons, 
 			 int score, int credits, PhysicsBodyDefinition bodyDefinition,
 
 			Listener<Integer> scoreListener) {
