+++ Bullet-Inferno/src/se/dat255/bulletinferno/model/entity/DefaultEnemyShipImpl.java
@@ -15,11 +15,9 @@ public class DefaultEnemyShipImpl extends SimpleEnemy implements Ship, Timerable
 	private Timer[] timers;
 	
 	public DefaultEnemyShipImpl(PhysicsEnvironment physics, EntityEnvironment entities,
-
-			EnemyDefinitionImpl type, Vector2 position, Vector2 velocity, int initialHealth,
-
-			Weapon[] weapons,
-			int score, int credits, PhysicsBodyDefinition bodyDefinition, Listener<Integer> scoreListener) {
+			EnemyDefinitionImpl type, Vector2 position, Vector2 velocity, float initialHealth,
+			Weapon[] weapons, int score, int credits, PhysicsBodyDefinition bodyDefinition, 
+			Listener<Integer> scoreListener) {
 
 		super(physics, entities, type, position, velocity, initialHealth, weapons,
 				 score, credits,
@@ -35,9 +33,7 @@ public class DefaultEnemyShipImpl extends SimpleEnemy implements Ship, Timerable
 	}
 	
 	public DefaultEnemyShipImpl(PhysicsEnvironment physics, EntityEnvironment entities,
-
-			EnemyDefinitionImpl type, Vector2 position, Vector2 velocity, int initialHealth,
-			
+			EnemyDefinitionImpl type, Vector2 position, Vector2 velocity, float initialHealth,
 			Weapon[] weapons, int score, int credits, PhysicsBodyDefinition bodyDefinition,
 			PhysicsMovementPattern pattern, Listener<Integer> scoreListener) {
 		super(physics, entities, type, position, velocity, initialHealth, weapons, score, credits,
